requirements.txt
paramiko==2.1.2
requests==2.13.0
PyYAML==3.12